+++ src/player.h
@@ -51,10 +51,10 @@ enum PlrTargetType {
 };
 
 //TODO would make more sense to have a struct for each agent, and then a merging struct
-struct AgentInfo {
-    ulong Weapons[CRYO_PODS_MAX_COUNT];
+struct AgentInfo { // sizeof=0x185
+    u32 Weapons[CRYO_PODS_MAX_COUNT];
     union Mod Mods[CRYO_PODS_MAX_COUNT];
-    ulong Sex;
+    u32 Sex;
     char RandomName[CRYO_PODS_MAX_COUNT];
     struct WeaponsFourPack FourPacks[CRYO_PODS_MAX_COUNT];
     ubyte NumAgents;
@@ -88,7 +88,7 @@ typedef struct {
     /** Double mode is multiple players using the same computer (controlling individual agents). */
     ubyte DoubleMode;
     ubyte PlayerNo;
-    ulong Weapons[AGENTS_SQUAD_MAX_COUNT];
+    u32 Weapons[AGENTS_SQUAD_MAX_COUNT];
     union Mod Mods[AGENTS_SQUAD_MAX_COUNT];
     /** Per-user controlled agent command param value Y */
     short UserVY[LOCAL_USERS_MAX_COUNT];
@@ -137,7 +137,7 @@ TbBool player_cryo_remove_weapon_one(ushort cryo_no, WeaponType wtype);
 TbBool player_cryo_transfer_weapon_between_agents(ushort from_cryo_no,
   ushort to_cryo_no, ubyte weapon);
 TbBool player_cryo_add_cybmod(ushort cryo_no, ubyte cybmod);
-const char *get_cryo_agent_name(ushort cryo_no);
+const char *get_cryo_agent_name(short cryo_no);
 void remove_agent(ubyte cryo_no);
 void add_agent(ulong weapons, ushort mods);
